WIP: ENH: fft: array API support [design option 4]
