+++ scipy/fft/__init__.py
@@ -89,10 +89,9 @@ from ._basic import (
 from ._realtransforms import dct, idct, dst, idst, dctn, idctn, dstn, idstn
 from ._fftlog import fhtoffset
 from ._fftlog_multimethods import fht, ifht
-from ._helper import next_fast_len
+from ._helper import next_fast_len, fftfreq, rfftfreq, fftshift, ifftshift
 from ._backend import (set_backend, skip_backend, set_global_backend,
                        register_backend)
-from numpy.fft import fftfreq, rfftfreq, fftshift, ifftshift
 from ._pocketfft.helper import set_workers, get_workers
 
 __all__ = [
