+++ scipy/fft/meson.build
@@ -2,11 +2,15 @@ python_sources = [
   '__init__.py',
   '_backend.py',
   '_basic.py',
+  '_basic_np.py',
   '_debug_backends.py',
   '_fftlog.py',
+  '_fftlog_np.py',
   '_fftlog_multimethods.py',
   '_helper.py',
-  '_realtransforms.py'
+  '_helper_np.py',
+  '_realtransforms.py',
+  '_realtransforms_np.py'
 ]
